Move functions back to onnxops
